chore: finish session info refactor
