+++ app/common/renderer/components/SessionInspector/SessionInfoTab/SessionInfoCodeBox.jsx
@@ -1,12 +1,12 @@
 import {Refractor} from 'react-refractor';
 
 import {CLIENT_FRAMEWORK_MAP} from '../../../lib/client-frameworks/map.js';
-import SessionCodeBoxCard from './SessionCodeBoxCard.jsx';
+import SessionInfoCodeBoxCard from './SessionInfoCodeBoxCard.jsx';
 
 /**
  * Code box with boilerplate code to start a session matching the current one.
  */
-const SessionCodeBox = ({clientFramework, setClientFramework, serverDetails, sessionCaps}) => {
+const SessionInfoCodeBox = ({clientFramework, setClientFramework, serverDetails, sessionCaps}) => {
   const {serverUrl, serverUrlParts} = serverDetails;
   const ClientFrameworkClass = CLIENT_FRAMEWORK_MAP[clientFramework];
 
@@ -14,14 +14,14 @@ const SessionCodeBox = ({clientFramework, setClientFramework, serverDetails, ses
   const clientCode = framework.getCodeString(true);
 
   return (
-    <SessionCodeBoxCard
+    <SessionInfoCodeBoxCard
       clientCode={clientCode}
       clientFramework={clientFramework}
       setClientFramework={setClientFramework}
     >
       <Refractor language={ClientFrameworkClass.refractorLang} value={clientCode} />
-    </SessionCodeBoxCard>
+    </SessionInfoCodeBoxCard>
   );
 };
 
-export default SessionCodeBox;
+export default SessionInfoCodeBox;
